refactor rename of data folder to db
