+++ src/pages/Work.js
@@ -14,7 +14,7 @@ import React from 'react';
 import { useParams } from 'react-router-dom'
 import BackBtn from '../components/BackBtn';
 import NotFound from '../pages/NotFound';
-const devData = require('../data/devWorks.json');
+const devData = require('../db/devWorks.json');
 
 // PATH: /dev/:workId
